Separate mkldnn quantizer test
